backend/requirements.txt
aiofiles==22.1.0
alembic==1.9.0
anyio==3.6.2
attrs==22.1.0
bcrypt==4.0.1
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
distlib==0.3.6
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.0.4
fastapi==0.88.0
filelock==3.9.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
identify==2.5.12
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
iniconfig==1.1.1
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
nodeenv==1.7.0
packaging==22.0
passlib==1.7.4
platformdirs==2.6.2
pluggy==1.0.0
pre-commit==2.21.0
psycopg2==2.9.5
pydantic==1.10.2
pytest==7.2.0
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.17.1
watchfiles==0.18.1
websockets==10.4
zipp==3.11.0
